refactor(syntax-transformer): undo rename
