+++ test/unit/ioc/syntax-transformer.spec.ts
@@ -1,12 +1,12 @@
 import * as ts from 'typescript';
 import { raw } from '../../fixture/ts/raw';
 import * as AST from '../../../ioc/analysis/ast';
-import { TypeScriptSyntaxTransformer } from '../../../ioc/analysis/ts-syntax-transformer';
+import { SyntaxTransformer } from '../../../ioc/analysis/syntax-transformer';
 
 describe('SyntaxTransformer', () => {
   let createSourceFileCount = 1; // Number.MAX_SAFE_INTEGER // restrict source files to process because it's cpu-intensive and clogs the console output
   let sourceFiles: ts.SourceFile[];
-  let sut: TypeScriptSyntaxTransformer;
+  let sut: SyntaxTransformer;
 
   beforeAll(() => {
     sourceFiles = [];
@@ -18,10 +18,10 @@ describe('SyntaxTransformer', () => {
       const sourceFile = ts.createSourceFile(path, raw[path], ts.ScriptTarget.ES2015);
       sourceFiles.push(sourceFile);
     }
-    sut = new TypeScriptSyntaxTransformer();
+    sut = new SyntaxTransformer();
   });
 
-  it('should transform to ioc AST', () => {
+  it('should transform typescript sources to ioc AST', () => {
     const config = sut.create(...sourceFiles);
 
     console.log(AST.toJSON(config as any));
